Implement new-style grouping for form
